refactor: uninstall component of kubesphere
