+++ kubectl-plugin/component/component/networkpolicy.go
