+++ kubectl-plugin/component/component/openpitrix.go
